[refactor] app, shared 레이어 추가
