+++ src/main.tsx
@@ -4,7 +4,7 @@ import { createRoot } from 'react-dom/client';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 
-import '@/styles/global.css';
+import '@/app/styles/global.css';
 
 import App from './App';
