+++ src/shared/utils/useOutsideClick.ts
